+++ programmable_video/test/mock_platform_interface.dart
@@ -12,6 +12,8 @@ class MockInterface extends ProgrammableVideoPlatform {
   var setSpeakerPhoneOnWasCalled = false;
   var speakerPhoneOn;
   var getSpeakerPhoneOnWasCalled = false;
+  var deviceHasReceiverWasCalled = false;
+  var getStatsWasCalled = false;
   var connectToRoomWasCalled = false;
   var enableAudioTrackWasCalled = false;
   var enableVideoTrackWasCalled = false;
@@ -73,6 +75,18 @@ class MockInterface extends ProgrammableVideoPlatform {
   Future<bool> getSpeakerphoneOn() {
     getSpeakerPhoneOnWasCalled = true;
     return Future.delayed(Duration(milliseconds: 1), () => speakerPhoneOn);
+  }
+
+  @override
+  Future<bool> deviceHasReceiver() {
+    deviceHasReceiverWasCalled = true;
+    return Future.delayed(Duration(milliseconds: 1), () => true);
+  }
+
+  @override
+  Future<Map<dynamic, dynamic>> getStats() {
+    getStatsWasCalled = true;
+    return Future.delayed(Duration(milliseconds: 1), () => {});
   }
 
   @override